created new folders and moved files around. website currently isn't working
