+++ src/scripts/data/dataManager.js
@@ -1,0 +1,8 @@
+export const getJournalEntries = () => {
+    const sortedByDate = journalEntries.sort(
+        (currentEntry, nextEntry) =>
+            Date.parse(currentEntry.date) - Date.parse(nextEntry.date)
+    )
+    return sortedByDate
+}
+// console.log(getJournalEntries());
